pyproject.toml
[tool.poetry]
name = "tax-regimes"
version = "0.0.1"
description = "Simple comparator for Indian tax regimes"
authors = ["Praseetha-KR <praseetha04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
texttable = "^1.6.2"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
flake8 = "^3.7.9"
ipython = "^7.13.0"
ipdb = "^0.13.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
